# Backend/openapi.yml
openapi: 3.0.0
info:
  title: Beacon API
  version: 1.0.0
  description: |
    Backend API for Beacon.

    Most `/api/*` endpoints require a JWT in `Authorization: Bearer <token>`.
    Request validation is enforced via `express-openapi-validator`.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/pins:
    get:
      tags: [Pins]
      summary: Get all pins
      description: |
        Retrieves all pins from the database. Supports optional distance sort via
        `sort=distance&lat=&lng=` (Near Me priority). When authenticated, includes
        the requesting user's status per pin (`visited` or `wishlist`).
      operationId: getAllPins
      parameters:
        - in: query
          name: sort
          required: false
          schema:
            type: string
            enum: [recent, distance]
          description: Sort mode. `distance` requires lat/lng query params.
        - in: query
          name: lat
          required: false
          schema:
            type: number
            minimum: -90
            maximum: 90
          description: Latitude of origin for distance sort.
        - in: query
          name: lng
          required: false
          schema:
            type: number
            minimum: -180
            maximum: 180
          description: Longitude of origin for distance sort.
      responses:
        200:
          description: Successful response with array of pins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pin'
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags: [Pins]
      summary: Create a new pin
      description: Creates a new pin with the provided data
      operationId: createPin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - latitude
                - longitude
              properties:
                latitude:
                  type: number
                  minimum: -90
                  maximum: 90
                  description: Latitude coordinate of the pin
                  example: 40.7128
                longitude:
                  type: number
                  minimum: -180
                  maximum: 180
                  description: Longitude coordinate of the pin
                  example: -74.0060
                title:
                  type: string
                  maxLength: 200
                  description: Title of the pin
                  example: "Golden Gate Overlook"
                description:
                  type: string
                  maxLength: 500
                  description: Description/message content of the pin
                  nullable: true
                image:
                  type: string
                  maxLength: 2000
                  description: URL of the image
                  nullable: true
                address:
                  type: string
                  maxLength: 200
                  description: Address of the pin location
                  nullable: true
                tags:
                  oneOf:
                    - type: string
                    - type: array
                      items:
                        type: string
                  description: Tags for the pin

      responses:
        201:
          description: Pin created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: ID of the newly created pin
        400:
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pins/trending:
    get:
      tags: [Pins]
      summary: Get trending pins
      description: |
        Returns up to 20 pins ordered by a trending score of
        `likes + 3 * max(0, 1 - ageDays / days)`. Recency score decays linearly
        and pins outside the window receive a recency contribution of 0.
      operationId: getTrendingPins
      parameters:
        - in: query
          name: days
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 365
            default: 7
          description: Number of days the linear recency decay is stretched over.
      responses:
        200:
          description: Successful response with trending pins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pin'
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pins/{id}/status:
    put:
      tags: [Pins]
      summary: Set visited/wishlist status on a pin
      description: Idempotent upsert of the requesting user's status for a pin.
      operationId: setPinStatus
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the pin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [visited, wishlist]
      responses:
        200:
          description: Status set successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  pinID:
                    type: integer
                  status:
                    type: string
                    enum: [visited, wishlist]
        400:
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Pin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Pins]
      summary: Clear visited/wishlist status for a pin
      description: Removes the requesting user's status entry for a pin.
      operationId: deletePinStatus
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the pin
      responses:
        204:
          description: Status cleared
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Status not set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pin-status:
    get:
      tags: [Pins]
      summary: Get all of the requesting user's pin statuses
      description: Returns every `pin_status` row belonging to the current user.
      operationId: getUserPinStatuses
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    pinID:
                      type: integer
                    status:
                      type: string
                      enum: [visited, wishlist]
                    updatedAt:
                      type: string
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/search/history:
    get:
      tags: [Pins]
      summary: Get recent search history
      description: Returns up to 10 most recent search queries for the current user.
      operationId: getSearchHistory
      responses:
        200:
          description: Array of recent queries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchHistoryEntry'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [Pins]
      summary: Record a search query
      description: |
        Records a search query for the current user. Duplicate queries are
        collapsed (the older entry is deleted so the new one is freshest).
        Retention is capped at 50 entries per user.
      operationId: addSearchHistory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                  maxLength: 200
      responses:
        201:
          description: Entry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchHistoryEntry'
        400:
          description: Invalid query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Pins]
      summary: Clear all search history
      description: Deletes every search history entry belonging to the current user.
      operationId: clearSearchHistory
      responses:
        204:
          description: History cleared
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/search/history/{id}:
    delete:
      tags: [Pins]
      summary: Delete a single search history entry
      operationId: deleteSearchHistoryEntry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: History entry ID
      responses:
        204:
          description: Entry deleted
        400:
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pins/user:
    get:
      tags: [Pins]
      summary: Get all pins created by the logged-in user
      description: Retrieves all pins created by the authenticated user
      operationId: getUserPins
      responses:
        200:
          description: Successful response with array of user's pins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pin'
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pins/nearby:
    post:
      tags: [Pins]
      summary: Get pins near a coordinate
      description: Returns all pins within 10 km of the given latitude/longitude, sorted by distance (closest first), using the Haversine formula for accurate great-circle distance.
      operationId: getPinsNearCoordinate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - latitude
                - longitude
              properties:
                latitude:
                  type: number
                  minimum: -90
                  maximum: 90
                  description: Latitude of the search origin
                  example: 37.7749
                longitude:
                  type: number
                  minimum: -180
                  maximum: 180
                  description: Longitude of the search origin
                  example: -122.4194
      responses:
        200:
          description: Array of pins within 10 km, sorted nearest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pin'
        400:
          description: Bad request - missing or invalid coordinates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pins/{id}:
    get:
      tags: [Pins]
      summary: Get a specific pin by ID
      description: Retrieves a single pin by its ID
      operationId: getPin
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the pin to retrieve
          example: 1
      responses:
        200:
          description: Successful response with pin data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pin'
        400:
          description: Bad request (invalid path parameter)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags: [Pins]
      summary: Update a specific pin by ID
      description: Updates title/message/image/address for a pin (owner only). Returns 404 for both missing pins and pins owned by another user to prevent ID enumeration.
      operationId: updatePin
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the pin to update

      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  maxLength: 200
                  nullable: true
                description:
                  type: string
                  maxLength: 500
                  nullable: true
                image:
                  type: string
                  maxLength: 2000
                  nullable: true
                address:
                  type: string
                  maxLength: 200
                  nullable: true

      responses:
        200:
          description: Pin updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pin'
        400:
          description: Bad request (invalid body or path parameter)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Unauthorized - not the owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Pin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Pins]
      summary: Delete a specific pin by ID
      description: Delete a single pin by its ID (owner only)
      operationId: deletePin
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the pin to delete
      responses:
        200:
          description: Successfully deleted pin
          content: {}
        400:
          description: Bad request (invalid path parameter)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Unauthorized - not the owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Pin not found
          content: {}
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pins/{pinId}/comments:
    get:
      tags: [Comments]
      summary: Get all comments for a pin
      description: Retrieves all comments for a specific pin
      operationId: getPinComments
      parameters:
        - in: path
          name: pinId
          required: true
          schema:
            type: integer
          description: The ID of the pin
      responses:
        200:
          description: Successful response with array of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags: [Comments]
      summary: Create a comment on a pin
      description: Posts a new comment on a specific pin
      operationId: createComment
      parameters:
        - in: path
          name: pinId
          required: true
          schema:
            type: integer
          description: The ID of the pin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - comment
              properties:
                comment:
                  type: string
                  maxLength: 280
                  description: The comment text
      responses:
        201:
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        400:
          description: Bad request - invalid comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Pin not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/comments/{commentId}:
    put:
      tags: [Comments]
      summary: Update a comment
      description: Updates a comment (owner only)
      operationId: updateComment
      parameters:
        - in: path
          name: commentId
          required: true
          schema:
            type: integer
          description: The ID of the comment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - comment
              properties:
                comment:
                  type: string
                  maxLength: 280
      responses:
        200:
          description: Comment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        400:
          description: Bad request - invalid comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Unauthorized - not the owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      tags: [Comments]
      summary: Delete a comment
      description: Deletes a comment (owner only)
      operationId: deleteComment
      parameters:
        - in: path
          name: commentId
          required: true
          schema:
            type: integer
          description: The ID of the comment
      responses:
        200:
          description: Comment deleted successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
                    example: "Comment deleted successfully"
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Unauthorized - not the owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Failed to delete comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/comments/{id}/reactions/{emoji}:
    delete:
      tags: [Comments]
      summary: Remove emoji reaction from comment
      description: Removes an emoji reaction from a comment
      operationId: removeCommentReaction
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the comment
        - in: path
          name: emoji
          required: true
          schema:
            type: string
          description: The emoji to remove
      responses:
        200:
          description: Reaction removed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Reaction removed"
        401:
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Reaction not found or comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SearchHistoryEntry:
      type: object
      properties:
        id:
          type: integer
        query:
          type: string
          maxLength: 200
        createdAt:
          type: string
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: Human-readable error message
          example: "Invalid token"
        error:
          oneOf:
            - type: string
            - type: object
          description: Error identifier/message (some endpoints return 'error' instead of 'message')
          example: "Failed to share itinerary"
        errors:
          type: array
          description: Detailed validation errors (OpenAPI validator)
          items:
            type: object
        details:
          type: string
          description: Additional error details
          example: "FOREIGN KEY constraint failed"
...